+++ backend/app/routers/users.py
@@ -1,0 +1,19 @@
+"""User management routes for the application."""
+from fastapi import APIRouter
+from app.dependencies import graph_client
+
+router = APIRouter(
+    prefix="/users",
+    tags=["users"],
+    # dependencies=[Depends(get_token_header)], # Uncomment if you want to enforce token validation for all user routes
+    responses={404: {"description": "Not found"}},
+)
+
+@router.get("/display-names")
+async def get_all_users_display_names():
+    """Endpoint to retrieve all users in the tenant."""
+    try:
+        result = await graph_client.users.get()
+        return {"users": [user.display_name for user in result.value]}
+    except (ValueError, AttributeError, TypeError) as e:
+        return {"error": str(e)}
